Started adding basic Minecraft functionality and came across an interesting 'feature'?
In Firefox (on Mac Mojave), this code causes a print prompt (naturally!). On Chrome, it seems to work as expected.

List<>.Append() wasn't working in Firefox (the list remained empty) so I couldn't write a GetNeighbours function to use for calculating the bomb-count that should be shown in a tile. Since normally zero-valued tiles cause all neighbouring tiles to be revealed, I switched to passing the Action to the method (either increment count to determine the displayed value, or call the neighbouring tile's reveal method), which causes the rather interesting behaviour described above.