Fixes the printer dialog error in Firefox (hopefully!) and makes the buttons visible in Chrome.
I'm still not sure why the printer dialog error occurs in Firefox but in Chrome, it reports "RangeError: Maximum call stack size exceeded" instead, so it seems a stack overflow occurs when revealing neighbours of 0-count buttons. Adding a variable limit that changes each call seems to fix the problem, but it seems like the value is ignored and the calls are just dropped whenever it reaches the limit (ie, setting max calls to 10 or 1000 has exactly the same behaviour - some neighbours aren't revealed that should be but no obvious error occurs).

A better solution might be rewriting this as a breadth-first rather than depth-first traversal, but that seems less readable so I haven't tried yet - and at this stage, I'd rather start implementing winning and losing behaviour.